Move function to bgp_helpers
